test: use local compiler path in `ExampleCommand` under tests
