+++ examples/fibonacci/cargo-generate.toml
@@ -1,0 +1,2 @@
+[template]
+ignore = ["target"]
