+++ tools/cargo-miden/tests/build.rs
@@ -33,7 +33,7 @@ fn new_project_args(project_name: &str, template: &str) -> Vec<String> {
 // that depend on the current directory
 
 #[test]
-fn test_all_templates() {
+fn test_all_templates_and_examples() {
     let _ = env_logger::Builder::from_env("MIDENC_TRACE")
         .is_test(true)
         .format_timestamp(None)
